refactor(webfont-generator): extract CSS hashing
