+++ packages/webfont-generator/src/rendering/css/mod.rs
@@ -1,5 +1,6 @@
 mod context;
 mod dependencies;
+mod hash;
 mod render;
 
 #[cfg(feature = "napi")]
@@ -12,7 +13,9 @@ pub(super) use render::render_css_with_context;
 pub(crate) use render::{render_css_with_hbs_context, render_css_with_src_mutate};
 
 #[cfg(test)]
-use context::{calc_hash, make_ctx, make_urls};
+use context::{make_ctx, make_urls};
+#[cfg(test)]
+use hash::calc_hash;
 #[cfg(test)]
 use render::render_default_css;
